requirements.txt
Flask==3.1.0
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.1.0
scikit_learn==1.6.1
Werkzeug==3.1.3
